Add multiple gamemode selectors
