+++ Utilla/GamemodeManager.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.UI;
 using System.Linq;
 using GorillaNetworking;
 using BepInEx;
@@ -9,59 +8,31 @@ using System.Reflection;
 using System.Linq.Expressions;
 using Photon.Pun;
 using Utilla.Models;
-using ExitGames.Client.Photon;
-using Photon.Realtime;
 
 namespace Utilla
 {
-	public partial class GamemodeManager : MonoBehaviour
+	public class GamemodeManager : MonoBehaviour
 	{
 		public static GamemodeManager Instance { get; private set; }
 
-		const string ModdedDefault = "MODDED_DEFAULT";
-		const string ModdedCasual = "MODDED_CASUAL";
-		const string ModdedHunt = "MODDED_HUNT";
-		const int PageSize = 3;
-
-		const string UIRootPath = "Level/forest/lower level/UI";
-		private const string AnchorPath = "Selector Buttons/anchor";
-		private const string GamemodesListPath = "Tree Room Texts/Game Mode List Text";
 
 		const string BasePrefabPath = "CustomGameManager/";
 
-		int page;
-		int PageCount => Mathf.CeilToInt(gamemodes.Count() / 3f);
+		public int PageCount => Mathf.CeilToInt(Gamemodes.Count() / 3f);
 
-		Text gamemodesText;
 		List<Gamemode> DefaultModdedGamemodes = new List<Gamemode>()
 		{
-			new Gamemode(ModdedCasual, "MODDED CASUAL", BaseGamemode.Casual),
-			new Gamemode(ModdedDefault, "MODDED", BaseGamemode.Infection),
-			new Gamemode(ModdedHunt, "MODDED HUNT", BaseGamemode.Hunt)
+			new Gamemode("MODDED_CASUAL", "MODDED CASUAL", BaseGamemode.Casual),
+			new Gamemode("MODDED_DEFAULT", "MODDED", BaseGamemode.Infection),
+			new Gamemode("MODDED_HUNT", "MODDED HUNT", BaseGamemode.Hunt)
 		};
-		List<Gamemode> gamemodes = new List<Gamemode>() { 
+		public List<Gamemode> Gamemodes { get; private set; } = new List<Gamemode>() { 
 			new Gamemode("CASUAL", "CASUAL"),
 			new Gamemode("INFECTION", "INFECTION"),
 			new Gamemode("HUNT", "HUNT")
 		};
-		ModeSelectButton[] modeSelectButtons = Array.Empty<ModeSelectButton>();
 
-		Material buttonMaterial = Resources.Load<Material>("objects/treeroom/materials/plastic");
-
-		public class PluginInfo
-		{
-			public BaseUnityPlugin Plugin { get; set; }
-			public Gamemode[] Gamemodes { get; set; }
-			public Action<string> OnGamemodeJoin { get; set; }
-			public Action<string> OnGamemodeLeave { get; set; }
-
-			public override string ToString()
-			{
-				return $"{Plugin.Info.Metadata.Name} [{string.Join(", ", Gamemodes.Select(x => x.DisplayName))}]";
-			}
-		}
-
-		List<PluginInfo> pluginInfos = new List<PluginInfo>();
+		List<PluginInfo> pluginInfos;
 
 		void Awake()
 		{
@@ -69,32 +40,49 @@ namespace Utilla
 			Events.RoomJoined += OnRoomJoin;
 			Events.RoomLeft += OnRoomLeft;
 
-			transform.parent = GameObject.Find(UIRootPath).transform;
+			// transform.parent = GameObject.Find(UIRootPath).transform;
 
             GorillaComputer.instance.currentGameMode = PlayerPrefs.GetString("currentGameMode", "INFECTION");
 
-			CreatePluginInfos();
+			pluginInfos = GetPluginInfos();
 
-			Transform anchor = transform.parent.Find(AnchorPath);
-			Transform[] buttons = Enumerable.Range(0, PageSize).Select(x => anchor.GetChild(x)).ToArray();
-			modeSelectButtons = buttons.Select(x => x.GetComponent<ModeSelectButton>()).ToArray();
+			Gamemodes.AddRange(GetGamemodes(pluginInfos));
+			Gamemodes.ForEach(gamemode => AddGamemodeToPrefabPool(gamemode));
 
-			CreateGamemodes();
-			CreatePrefabPool();
+			try
+			{
+				var treehouseSelector = new GameObject("TreehouseSelector").AddComponent<GamemodeSelector>();
+								Transform treehouseParent = GameObject.Find("Level/forest/lower level/UI").transform;
+				treehouseSelector.Initialize(treehouseParent,
+						   treehouseParent.Find("Selector Buttons/anchor"),
+						   treehouseParent.Find("Tree Room Texts/Game Mode List Text"));
+			}
+			catch (Exception e)
+			{
+				Debug.LogError($"Utilla: Failed to initialize TreehouseSelector: {e}");
+			}
 
-			CreateGamemodeText();
-
-			CreatePageButtons(buttons[0].gameObject);
-
-			ShowPage(0);
+			try
+			{
+				var mountainSelector = new GameObject("MountainSelector").AddComponent<GamemodeSelector>();
+				Transform mountainParent = GameObject.Find("Level/mountain").transform;
+				mountainSelector.Initialize(mountainParent,
+						   mountainParent.Find("Geometry/goodigloo/modeselectbox (1)/anchor"),
+						   mountainParent.Find("UI/Text/Game Mode List Text (1)"));
+			}
+			catch (Exception e)
+			{
+				Debug.LogError($"Utilla: Failed to initialize MountainSelector: {e}");
+			}
 		}
 
-		void CreateGamemodes()
+		List<Gamemode> GetGamemodes(List<PluginInfo> infos)
 		{
+			List<Gamemode> gamemodes = new List<Gamemode>();
 			gamemodes.AddRange(DefaultModdedGamemodes);
 
 			HashSet<Gamemode> additonalGamemodes = new HashSet<Gamemode>();
-			foreach (var info in pluginInfos)
+			foreach (var info in infos)
 			{
 				additonalGamemodes.UnionWith(info.Gamemodes);
 			}
@@ -105,57 +93,13 @@ namespace Utilla
 			}
 
 			gamemodes.AddRange(additonalGamemodes);
+
+			return gamemodes;
 		}
 
-		void CreateGamemodeText()
+		List<PluginInfo> GetPluginInfos()
 		{
-			gamemodesText = transform.parent.Find(GamemodesListPath).gameObject.GetComponent<Text>();
-			gamemodesText.lineSpacing = 2.2f;
-			gamemodesText.transform.localScale *= 0.75f;
-			gamemodesText.transform.position += gamemodesText.transform.right * 0.1f;
-			gamemodesText.horizontalOverflow = HorizontalWrapMode.Overflow;
-		}
-
-		void CreatePageButtons(GameObject templateButton)
-		{
-			GameObject cube = GameObject.CreatePrimitive(PrimitiveType.Cube);
-			cube.SetActive(false);
-			MeshFilter meshFilter = cube.GetComponent<MeshFilter>();
-
-			GameObject CreatePageButton(string text, Action onPressed)
-			{
-				GameObject button = GameObject.Instantiate(templateButton);
-				button.GetComponent<MeshFilter>().mesh = meshFilter.mesh;
-				button.GetComponent<Renderer>().material = buttonMaterial;
-				button.transform.parent = templateButton.transform.parent;
-				button.transform.localRotation = templateButton.transform.localRotation;
-				button.transform.localScale = templateButton.transform.localScale;
-
-				button.transform.GetChild(0).gameObject.SetActive(true);
-				Text buttonText = button.GetComponentInChildren<Text>();
-				if (buttonText != null)
-				{
-					buttonText.text = text;
-					buttonText.transform.localScale = Vector3.Scale(buttonText.transform.localScale, new Vector3(2, 2, 1));
-				}
-
-				GameObject.Destroy(button.GetComponent<ModeSelectButton>());
-				button.AddComponent<PageButton>().onPressed += onPressed;
-
-				return button;
-			}
-
-			GameObject nextPageButton = CreatePageButton("-->", NextPage);
-			nextPageButton.transform.localPosition = new Vector3(-0.575f, nextPageButton.transform.position.y, nextPageButton.transform.position.z);
-
-			GameObject previousPageButton = CreatePageButton("<--", PreviousPage);
-			previousPageButton.transform.localPosition = new Vector3(-0.575f, -0.318f, previousPageButton.transform.position.z);
-
-			Destroy(cube);
-		}
-
-		void CreatePluginInfos()
-		{
+			List<PluginInfo> infos = new List<PluginInfo>();
 			foreach (var info in BepInEx.Bootstrap.Chainloader.PluginInfos)
 			{
 				if (info.Value == null) continue;
@@ -167,7 +111,7 @@ namespace Utilla
 
 				if (gamemodes.Count() > 0)
 				{
-					pluginInfos.Add(new PluginInfo
+					infos.Add(new PluginInfo
 					{
 						Plugin = plugin,
 						Gamemodes = gamemodes.ToArray(),
@@ -176,6 +120,8 @@ namespace Utilla
 					});
 				}
 			}
+
+			return infos;
 		}
 
 		Action<string> CreateJoinLeaveAction(BaseUnityPlugin plugin, Type baseType, Type attribute)
@@ -232,34 +178,26 @@ namespace Utilla
 			return gamemodes;
 		}
 
-		void CreatePrefabPool()
+		void AddGamemodeToPrefabPool(Gamemode gamemode)
 		{
-			List<GameObject> prefabs = new List<GameObject>();
-			foreach (Gamemode gamemode in gamemodes)
-			{
-				if (gamemode.GameManager is null) continue;
+			if (gamemode.GameManager is null) return;
 
-				GameObject prefab = new GameObject(gamemode.ID);
-				prefab.SetActive(false);
-				prefab.AddComponent(gamemode.GameManager);
-				prefab.AddComponent<PhotonView>();
-				prefabs.Add(prefab);
-			}
+			GameObject prefab = new GameObject(gamemode.ID);
+			prefab.SetActive(false);
+			prefab.AddComponent(gamemode.GameManager);
+			prefab.AddComponent<PhotonView>();
 
 			DefaultPool pool = PhotonNetwork.PrefabPool as DefaultPool;
-			if (pool != null)
-			{
-				prefabs.ForEach(prefab => pool.ResourceCache.Add(BasePrefabPath + prefab.name, prefab));
-			}
+			pool.ResourceCache.Add(BasePrefabPath + prefab.name, prefab);
 		}
 
-		internal void OnRoomJoin(object sender, Events.RoomJoinedArgs args)
+		void OnRoomJoin(object sender, Events.RoomJoinedArgs args)
 		{
 			string gamemode = args.Gamemode;
 
 			if (PhotonNetwork.IsMasterClient)
 			{
-				foreach(Gamemode g in gamemodes.Where(x => x.GameManager != null))
+				foreach(Gamemode g in Gamemodes.Where(x => x.GameManager != null))
 				{
 					if (gamemode.Contains(g.ID))
 					{
@@ -286,7 +224,7 @@ namespace Utilla
 			}
 		}
 
-		internal void OnRoomLeft(object sender, Events.RoomJoinedArgs args)
+		void OnRoomLeft(object sender, Events.RoomJoinedArgs args)
 		{
 			string gamemode = args.Gamemode;
 
@@ -304,53 +242,6 @@ namespace Utilla
 					}
 				}
 			}
-		}
-
-		public void NextPage()
-		{
-			if (page < PageCount - 1)
-			{
-				ShowPage(page + 1);
-			}
-		}
-
-		public void PreviousPage()
-		{
-			if (page > 0)
-			{
-				ShowPage(page - 1);
-			}
-		}
-
-		void ShowPage(int page)
-		{
-			this.page = page;
-			List<Gamemode> currentGamemodes = gamemodes.Skip(page * PageSize).Take(PageSize).ToList();
-
-			int counter = 0;
-			for (int i = 0; i < modeSelectButtons.Length; i++)
-			{
-				if (i < currentGamemodes.Count)
-				{
-					modeSelectButtons[i].enabled = true;
-					modeSelectButtons[i].gameMode = currentGamemodes[i].GamemodeString;
-				}
-				else
-				{
-					modeSelectButtons[i].enabled = false;
-					modeSelectButtons[i].gameMode = "";
-					counter++;
-				}
-			}
-
-			string displayText = string.Join("\n", currentGamemodes.Select(x => x.DisplayName));
-			for (int i = 0; i < counter; i++)
-			{
-				displayText += '\n';
-			}
-			gamemodesText.text = displayText;
-
-			GorillaComputer.instance.OnModeSelectButtonPress(GorillaComputer.instance.currentGameMode);
 		}
 	}
 }
