+++ Utilla/PluginInfo.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Linq;
+using BepInEx;
+using Utilla.Models;
+
+namespace Utilla
+{
+	public class PluginInfo
+	{
+		public BaseUnityPlugin Plugin { get; set; }
+		public Gamemode[] Gamemodes { get; set; }
+		public Action<string> OnGamemodeJoin { get; set; }
+		public Action<string> OnGamemodeLeave { get; set; }
+
+		public override string ToString()
+		{
+			return $"{Plugin.Info.Metadata.Name} [{string.Join(", ", Gamemodes.Select(x => x.DisplayName))}]";
+		}
+	}
+}
